logic for tracking files to validate after checkout on NFS
Summary:
## Context
S439820 is an ongoing SEV that has been challenging to reproduce. Initially, we had some repros, but they suddenly disappeared. We suspect that NFS cache invalidation might be causing stale files on macOS. There is no direct way to invalidate cache in NFS. Therefore, we have to use some hacky ways to invalidate cache on macOS see this diff summary for detailed explanation D35435764

## This Diff Stack
This diff stack aims to implement an invalidation check method to verify the invalidation after each checkout on macOS. This will help us detect how many users are affected by the SEV and confirm whether our NFS cache invalidation approach is working correctly.
Also, if we find a large number of users are affected by the SEV, and we try any fix for macOS invalidation this check can confirm if any fix works in prod.

## This diff
Implements `maybeRecordInvalidation ` in the `CheckoutContext` then we can use it in later diffs to record some inodes for invalidation check

Reviewed By: lXXXw

Differential Revision: D68586870

fbshipit-source-id: 1b9d0a1e52da3f5420f79018ceef1b01df1157c2