Return invalidation list of inodes after checkout on NFS
Summary:
## Context
S439820 is an ongoing SEV that has been challenging to reproduce. Initially, we had some repros, but they suddenly disappeared. We suspect that NFS cache invalidation might be causing stale files on macOS. There is no direct way to invalidate cache in NFS. Therefore, we have to use some hacky ways to invalidate cache on macOS see this diff summary for detailed explanation D35435764

## This Diff Stack
This diff stack aims to implement an invalidation check method to verify the invalidation after each checkout on macOS. This will help us detect how many users are affected by the SEV and confirm whether our NFS cache invalidation approach is working correctly.
Also, if we find a large number of users are affected by the SEV, and we try any fix for macOS invalidation this check can confirm if any fix works in prod.

## This diff
Finaly `CheckoutContext::finish()` returns the invalidation list along with the checkout conflicts in a pair to process in the next diff.

Reviewed By: jdelliot, genevievehelsel

Differential Revision: D68631239

fbshipit-source-id: 0d17e7af8b3e202d744dabb7eaa3e2b4d03df84d