add output to 'eden prefetch-profile fetch-predictive'
Summary:
# This diff

Adds some informative output to 'eden prefetch-profile fetch-predictive' so that users know that the command was successful.

# Context

It's unclear to users that the `prefetch-profile fetch-predictive` command worked as expected. This is because no output is printed, and the verbose option doesn't even print out whether the prefetch was started/finished.

Adding this output should give users advice on how to check if the prefetch is doing work or not.

Reviewed By: genevievehelsel

Differential Revision: D68860827

fbshipit-source-id: 22355e525d51d9ba2b43cb1fdb2d01e2579d832b