more verbosity in 'eden prefetch-profile fetch-predictive'
Summary:
# This diff

Makes 'eden prefetch-profile fetch-predictive' more verbose by removing some verbosity restrictions. This makes the output of the command more informative.

# Context

It's unclear to users that the `prefetch-profile fetch-predictive` command worked as expected. This is because no output is printed, and the verbose option doesn't even print out whether the prefetch was started/finished.

Adding this output should give users advice on how to check if the prefetch is doing work or not.

Reviewed By: genevievehelsel

Differential Revision: D68860828

fbshipit-source-id: 80368d6dfaaa32b21072e5aeb5fe0f37862085e7