Fix context section misalignment
Summary:
Clicking "Expand 35 lines" between sections of the comparison view was broken in some comparions.

The convention we were using is to fetch the context lines in the "before" side of the comparison. The lines are the same in the "after" side, the only difference is the line numbers.

The bug here is that in committed comparisons, we were using `sl cat --rev hash filepath`, which means we were getting the "after" side of the file contents. But we actually want the "before" side. This was not a bug for uncommitted/head changes, where we use `sl cat --rev .` and sl cat --rev .^`, which is for the before side (the file contents on disk are the "after" side here).

I think we may need to audit some other uses of `sl cat`, since this is not the only use of the `revsetForComparison` util. But for now this diff focuses on the context lines fetching.

Reviewed By: quark-zju

Differential Revision: D68918165

fbshipit-source-id: 858b576ff0fe241cf50d8b1d8cd91397b29f641e