edenapi: rename MinTransferSpeed grace_period -> window
Summary:
"Grace period" suggests the limit does not apply until after the period, but that isn't how it works. The duration is the window over which the speed is calculated (and as a side-effect of that, the limit can't be applied until after the first window has completed).

I also renamed the config knob to "edenapi.low-speed-window", keeping support for the existing "edenapi.low-speed-grace-period-seconds".

Reviewed By: lmvasquezg

Differential Revision: D68860537

fbshipit-source-id: ab47268a719f327563030ddcf173d445b339a4f8