cat: remove single file optimization
Summary:
With the advent of the tree manifest, I don't think this optimization makes sense. Either way you only need to load the ancestor trees of the file path you are looking up.

This optimization was originally added in 46ead70cdff366faab1609bce666b8da77565e56 with this message:

    Special case the single file case in hg cat. This allows us to avoid
    parsing the manifest, which shaves 15% off hg cat perf. This is worth
    it, since automation often uses hg cat for retrieving single files.

"Parsing the manifest" refers to parsing the entire flat manifest to perform the walk operation, but walk is naturally optimized now via the tree manifest.

Reviewed By: MichaelCuevas

Differential Revision: D68860535

fbshipit-source-id: 5f477c332c24eb0b821313690536ab41de404a20