edenapi: separate out timeout for opening connection
Summary:
Add separate config knob corresponding to curl's CURLOPT_CONNECTTIMEOUT. This is a timeout for the initial DNS lookup and TCP/TLS initialization. Previously it was using the default 300s, I think.

If this part gets stuck, we don't want to wait 5 minutes before retrying. Let's lower the default values to 5s on prod and 10s elsewhere.

Reviewed By: lmvasquezg

Differential Revision: D68860536

fbshipit-source-id: e30131cf629c3b1dfc8f56f64fd5bba6fc5c0a68