edenapi: make low speed check configurable per-handler
Summary:
Allow curl's "low speed" threshold to be configured per-SLAPI handler.

Curl's "low speed" detection is essentially a smarter timeout that works based on transfer speed. It allows for long requests as long as they are continuously streaming data (i.e. making progress). Previously it was configured globally, so it was basically useless since it had to be configured to handle the slowest/worst expected case.

Reviewed By: MichaelCuevas

Differential Revision: D68860532

fbshipit-source-id: e87745900f5ea01cac727696aa053e4afb0b67a9