wireproto: Don't depend on very specific slog idioms
Summary:
We want to migrate from `slog` to `tracing`.

Code that uses advanced `slog` features such as customizing Loggers to duplex information across multiple drains and setting KV filters makes that more difficult.

Instead, use much simpler rust to achieve the intended effect:
In some cases, we want to send the errors back to the client by writing it to a mpsc channel of Bytes.

Reviewed By: markbt

Differential Revision: D68902095

fbshipit-source-id: 378336b7c2d13f7fb0f8a8027b181487992228fd