Add java-swift language support
Summary: This is a similar hack to what buck did with the javadeprecated client, adding a java_binary so we can get a self-contained set of fbcode deps over to fbandroid as a jar file. I don't love it and would be happy if someone knows an alternative. Moving to xplat was one failed approach I tried. It failed because the swift runtime still requires fbcode's third-party-java.

Reviewed By: cute-jumper

Differential Revision: D68968874

fbshipit-source-id: 1337a92b05929c331c6986fab582a40bad2146c8