Set the log level for the logging library
Summary:
We've been setting the `glog` level (used by C++ libraries) to match our application's log level (`slog`, currently). However, some of our dependencies still use the legacy `logging` library; the log level for that is set separately.

This diff implements a minimal fix. We will overhaul this anyway when moving from `slog` to `tracing`, and when we do we can make different choices. In the meantime this makes our library quieter and unblocks the ODS3 migration.

It is quite possible this will also reduce the console noise on AWS deployments, but I'll wait for a push to confirm.

Reviewed By: RajivTS

Differential Revision: D69003867

fbshipit-source-id: 225bcdff7f131984573f12cac1dbcd6f55b2b594