modern sync: fail faster
Summary: Modern sync was "running" but in reality was doing... nothing, given that the only time we await for changesets was when the bookmark moves, which... well... is not ideal cause contents and trees and files are getting synced but the changeset never really makes it.  And the metrics are not really accurate, so fixing that here.

Reviewed By: YousefSalama

Differential Revision: D69063466

fbshipit-source-id: 6ec862ea0de41aa2502e07851dce0b56b939d983