Introduce thirft struct for GitSymbolicRefs
Summary:
We plan to introduce caching for `GitSymbolicRefs` since its a mostly-static piece of data that is associated with each Git repo. It stores symrefs in a given Git repo (like `HEAD` for example). Most Git repos just have the `HEAD` symref which once created is rarely changed.

We need to access and query all known symrefs for a repo everytime we serve reads through Mononoke Git server. This results in a lot of unnecessary queries to the DB which can be prevented if we can just cache the results.

This diff is the first in the stack to achieving that objective

Reviewed By: clara-9

Differential Revision: D69178880

fbshipit-source-id: 7fb08af6fc956bc3adeaa866dd3c43b8becee984