repo_import: ensure topological order of existing commits is maintained
Summary: During repo_import, some commits may already be imported into the repo, so these can be skipped.  However, when doing this, we must still maintain the order, as we will be processing them in this order and we need to process them topologically (so that parents are processed before their children).

Reviewed By: giorgidze

Differential Revision: D69180445

fbshipit-source-id: 45daa6763e0afa619a51cbc91aed168516022c87