add telemetry for file util io retries
Summary:
# This diff

Adds 2 counters to track whether retried filesystem IO operations fail/succeed.

# Context

We'd like to track how often retried IO operations fail/succeed. We want to do this so that instead of just hiding IO failures with retries, we can actively work on driving them to 0.

If we log both when the IO retry fails/succeeds, we'll know how often retries are occurring. We can then work to drive the failure case down to 0, and eventually drive all retry cases to 0 as well.

Reviewed By: muirdm

Differential Revision: D69207098

fbshipit-source-id: 36b20684eeceedcede248d15f2cb34736eb0a213