move sys module to separate sysutil crate
Summary:
# This diff

Moves sys util functions to a separate sysutil crate

# Context

We'd like to track how often retried IO operations fail/succeed. We want to do this so that instead of just hiding IO failures with retries, we can actively work on driving them to 0.

To do this, we need to utilize the metrics crate which has a dependency on the utils. This introduces a circular dependency. We can easily remove this by removing the sys util functions into a separate sysutil crate. This removes the `metrics` crate dependency on the wider `util` crate.

Reviewed By: muirdm

Differential Revision: D69207097

fbshipit-source-id: 9be16414ff56add73152e2ffd2f81399c5a46f74