Publish net_util timeseries
Summary:
This diff is a bug fix. We use the timeseries "mononoke.lfs.download.net_util" to keep track of bandwith utilisation. This is then used to load shed if we are close to reaching the upper limits:
https://www.internalfb.com/code/configerator/source/scm/mononoke/lfs_server/default.cinc?lines=17-22

https://www.internalfb.com/code/fbsource/fbcode/eden/mononoke/lfs_server/src/service/middleware.rs?lines=72-79

The limits also include earlier load shedding for certain non-interactive users such as mactest.

We accidentally stopped publishing this timeseries in D57861971 which was attempting to remove CSLB related code, which also used the timeseries (but was not the exclusive user). It went unnoticed because
1) the system was designed to fail open
2) due to fixes at around the same time increased capacity and routing improvements led to not hitting the threshold.

## Reoccurrence prevention

We should set an alarm for ODS timeseries used for ratelimiting and loadshedding

Reviewed By: andreacampi

Differential Revision: D69127041

fbshipit-source-id: f84d5adff2db41ad659b867d02375f0df31669c4