Add stub for eden debug subscribe - for help text
Summary:
# Context
In D68223853 we removed rust help text for eden debug and subcommands since it was overriding the python help text, which contained most of the debug commands. This had the fallout of removing the help text for rust-only subcommands of debug that didn't have a matching stub, aka subscribe

# This Diff
Add a stub for eden debug subscribe with the same arguments as the rust version. This stub raises NotImplemented if force called via SKIP_RUST

Reviewed By: genevievehelsel, lXXXw

Differential Revision: D69259712

fbshipit-source-id: 5f000e98bbf7091da8de0ca5ef95f5823f26f48f