posix: add fallback for cross-device rename() failure
Summary:
If `os.rename()` fails due to a cross-device link error, try the `shutil.move()` method, which handles this. The operation is no longer as efficient or atomic, but it should work.

I want this in particular for the next diff D68978188 which moves the ".hg" repo dir off of the eden mount. Thus, any `util.rename()` operations spanning the repo dir and the working copy will become cross-device (and some tests fail).

It could be dangerous to silently fall back to non-atomic functionality, but after auditing the uses of util.rename() it was non-trivial to separate the use cases. I don't expect it to matter in practice.

Reviewed By: zzl0

Differential Revision: D69147655

fbshipit-source-id: 82729b851fe295688be53dae353c41bae468a0e8