clone: support symlinking to off-mount repo dir
Summary:
If clone.off-mount-repo-dir is set (or we are in an integration test), create the working copy's repo dir (i.e. ".hg" dir) in the repo's state dir, and drop a symlink in the working copy.

The goal is to move the repo dir off the eden mount so EdenFS can be oblivious to the files in ".hg" (e.g. eden shouldn't need to manage the repo files in its overlay).

Reviewed By: sggutier

Differential Revision: D68978188

fbshipit-source-id: 27279b24e33cad16aa51db6078d12854a282cb7a