clone: add checkout config for off-mount-repo-dir
Summary:
This "clone.off-mount-repo-dir" config will control whether we create the mount's ".hg" directory outside the mount.

We persist the new config value to the checkout yaml, although I'm not sure if that is strictly necessary. It will at least be useful to authoritatvely tell if the config was enabled when a repo was cloned.

Reviewed By: MichaelCuevas

Differential Revision: D68978190

fbshipit-source-id: 699673a1e22533a34682fa46a0f1ab79589b5dc2