Add RE worker id to the entity string
Summary:
# Context
The REMOTE_EXECUTION_WORKER is an enviornmental variable which is set for RE workers in:
https://www.internalfb.com/code/fbsource/[a1068b16d3f8]/fbcode/one_world/resource_managers/scm_manager/edenfs_repo.py?lines=57
We need to check this variable and if it is set then add it to the entity list for the OBC counters. Therefore, we can run ODS query per hostname and find the counter value for all the RE workers id in a TW task.

For example if hostname  = "twshared36511.05.cco1" and worker_is is "25", then the entities should be {"twshared36511.05.cco1", "twshared36511.05.cco1:25"} Therefore the values can be queried by both hostname and the worker id inside the hostname.

This env is only set for RE jobs, therefroe this code doesn't have any effect on other platforms.

# This diff
Change the OBC bump for eden RSS memory OBC counter

Reviewed By: ViniGupta08

Differential Revision: D69146281

fbshipit-source-id: b118fe0e06d009607b5f69acfe9ca06f2d6aa573