Add option to print mapping of git commit id -> bonsai changeset id when importing repo
Summary:
Currently, there's no way to find the final commit of any bookmark besides master.

If we print the mapping of commits, we can do `original git commit -> bonsai in large repo -> rewritten bonsai in large repo`.
Then we can manually create branches on the large repo, if needed.

This is not the best solution, but it will work quickly.

Reviewed By: andreacampi

Differential Revision: D69249229

fbshipit-source-id: c91cb51bba716f6aa9be0f1f2abe1b5d6f765cc1