buffer: avoid aliasing MmapMut references (#1025)
Summary:
`Buffer` has an `UnsafeCell<MmapMut>`. It protects against concurrent writes to it by using a mutex. However, it has UB by allowing callers of `read()` and `write()` to concurrently have mutable and immutable references to the `MmapMut`. This patch avoids the alias by using a pointer instead.

Thanks to Ralith for spotting the problem and suggesting the fix during our review of unsafe Rust code at Google.

Pull Request resolved: https://github.com/facebook/sapling/pull/1025

Reviewed By: quark-zju

Differential Revision: D69315548

fbshipit-source-id: 59ec97932b4e7013f287d749a4ebe2f5c2aaf313