scsc: keep using tokio::spawn for scsc
Summary:
In D68960884 we switched to `mononoke::spawn_task` everywhere so that we can automatically wire up things like `tracing` spans.  However, `scsc` is not a normal mononoke process, and more importantly, is cross-platform, while the `mononoke_macros` crate is not.  This prevents `scsc` from building on other platforms like windows.

Fix this by reverting the change for `scsc`.  It doesn't use `tracing`, so this isn't a problem.

Reviewed By: lmvasquezg

Differential Revision: D69394572

fbshipit-source-id: 799b59a384857f1f1f4d4f9742dfc2f3d375bde4