Implement core GitTree derivation logic
Summary:
Using the manifest trait, this diff defines the core GitTree derivation logic based on changes from BonsaiChangeset. Note that even though I call this derivation logic, this won't be used to create and save a different derived data type. It will instead be derived as part of GitCommit derived data type when needed.

The next diff in the stack will tie all the pieces together and update the GitCommit derivation logic.

Reviewed By: markbt

Differential Revision: D69257146

fbshipit-source-id: 56d2617a4be0e85ec8ed86588c974dea3f6bb252