Implement support for generating GitLeaf from Mononoke commit file changes
Summary:
The `TreeHandle` datastructure was created back when we didn't have raw Git Trees and raw Git Commits in our Mononoke storage, but now we do. Mark's changes last year implemented the manifest trait for the raw Git Tree type.

With all this being done, the thrift Git Tree type (named as `TreeHandle`) in our code is a redundancy that needs to be eliminated. Infact, the whole Git Tree derived data type needs to be eliminated and the tree derivation can happen when we are deriving Git commit. In the case of Git push and GitImport, both Git Tree and Git Commit already exist so we would be able to do away with this unnecessary derivation on our end.

This is the first diff in the stack that does that.

Reviewed By: markbt

Differential Revision: D69253959

fbshipit-source-id: 28e92fe0a417b03706dc2f4cd085547598665ccc