Add RE worker id to the entity string for OBC counters
Summary:
# Context
The REMOTE_EXECUTION_WORKER is an enviornmental variable which is set for RE workers in:
https://www.internalfb.com/code/fbsource/[a1068b16d3f8]/fbcode/one_world/resource_managers/scm_manager/edenfs_repo.py?lines=57
We need to check this variable and if it is set then add it to the entity list for the OBC counters. Therefore, we can run ODS query per hostname and find the counter value for all the RE workers id in a TW task.

For example if hostname  = "twshared36511.05.cco1" and worker_is is "25", then the entities should be {"twshared36511.05.cco1", "twshared36511.05.cco1:25"} Therefore the values can be queried by both hostname and the worker id inside the hostname.

This env is only set for RE jobs, therefroe this code doesn't have any effect on other platforms.

# This diff
Add new entity to the OBC API for scm store counters

Reviewed By: liubov-dmitrieva

Differential Revision: D69277647

fbshipit-source-id: 976cb03caae34fc4e970d339ce17c43fb1958a1f