introduce a new type for fuse unmount options
Summary:
## This Diff

Introduce a new type for unmount (only for fuse as a starting point). This struct will be carrying options like "force". By default we are using "--force" so setting it to true.

## Context
Extend `eden rm` with a "--no-force" flag so "--force" is not used when unmounting a mount point.

Reviewed By: genevievehelsel

Differential Revision: D69149657

fbshipit-source-id: ccbae761195a5438428bed47464c1678f8a0ef7a