update privHelper client/server [de]serialization of UnmountOptions
Summary:
## This Diff
Actually start using the new type `UnmountOptions`. We modified privHelper Client to serialize the request to include new fields. We also modified privHelper server to deserialize the new fields.

Currently only "force" will be checked since we are only supporting "--no-force" for now.

## Context
Extend eden rm with a "--no-force" flag so "--force" is not used when unmounting a mount point.

Reviewed By: MichaelCuevas

Differential Revision: D69165546

fbshipit-source-id: 07198493fc883d8d356ae6822f850d5fc6efdb8c