Enhance shallow object information in Mononoke Git
Summary:
When we get rid of Tree Handles, we would need to serve shallow responses based on raw Git trees instead of thrift Tree Handles. This means working with Git commits instead of Bonsais. However, there are certain other contexts in which we do need Bonsais.

In this diff, I have extended the definition of the Shallow Response struct to include both bonsais and git object ids so both use cases can be satisfied.

Differential Revision: D69398543

fbshipit-source-id: 613ab303760e29b5e8fd5ef6567a8a284cd6fc60