Surface errors during reads in the user response from Mononoke Git
Summary:
In the current implementation, if there is any error while executing clone/pull/fetch from Mononoke Git server, the user just sees a `EOF error` or `Expected more data` error or some other variant which tells little to nothing of the actual cause of the problem.

This diff ensures that whenever we encounter such an error, we send the error response back to the client so it shows up on the user CLI. This will make debugging and oncall much more simpler.

Reviewed By: lmvasquezg

Differential Revision: D69456016

fbshipit-source-id: 12dca9d5fc6682678c6a2d31a05970a37c10e861