Implement shallow read path support without thrift Git trees
Summary: The shallow clone/fetch/pull was the last remnant of code that relied on thrift Git trees (i.e. TreeHandle) to fetch the necessary objects to populate the packfile. This diff makes the necessary changes to update the logic to instead rely on the native Git trees thus freeing us up to deprecate the thrift trees moving forward.

Differential Revision: D69413333

fbshipit-source-id: 5db309e3ff6d0b2e4c523eefb4968e8ac8e28587