Integration test highlighting lack of error context during Mononoke Git reads
Summary:
Currently, the Mononoke Git server returns meaningful error information only during pushes / writes. For writes, if the error happens before we start computing the response packfile, it shows up in a reasonable form at the client end.

However, the crux of the read logic in Mononoke Git resides in the packfile generation. In the current state, any such error results in the client seeing `early EOF` or `fetch-pack: invalid index-pack output` both of which are less than useful for the user and for the oncall trying to investigate the problem.

This diff adds an integration test that highlights that problem. I will fix this issue in the follow up diff.

Reviewed By: lmvasquezg

Differential Revision: D69455834

fbshipit-source-id: 4492021056421707ecbc9df9381f1a959ca3dbdb