Define new enum types for privHelperServer to handle requests to start/stop FAM
Summary:
## This Diff

Defined two new enum types to start/stop FAM. These types will be used by privHelperClient to serialize into the requests and by the privHelperServer to decide how to process the requests.

## Context
File Access Monitor(FAM) is a tool to monitor file system access, aiming to find out EdenFS crawlers. The tool will be implemented after the toolchain is ready. Before that, we are going to use [this third-party binary](https://github.com/objective-see/FileMonitor).
This diff stack focuses on integrating the binary into Eden package and spawn a process from privHelper because of the required privileged permission.

Reviewed By: ViniGupta08

Differential Revision: D69340299

fbshipit-source-id: 4cc8176162e7099e4263b9619de3d2a44a60a219