Log packfile errors to Mononoke Git scuba
Summary:
In D69456016, we enabled server side errors during packfile creation for serving read traffic in Mononoke Git to be surfaced to the user's git client. Although that does help in troubleshooting production issues, it would also be useful to have the same logs show in scuba so the oncall can investigate after the fact and monitor the trend of such errors over time.

This diff adds the necessary scuba logging

Reviewed By: andreacampi

Differential Revision: D69464304

fbshipit-source-id: 500da5848ea604aa67ebf2ad2a50d687534c9620