do not panic, but gracefully fail on non-40B ref contents. ignore symrefs
Summary:
The utility in monad that creates a git bundle from on-disk repo, assumes each ref is directly 40B sha1.

It does this while getting a slice, which in the case of indirect refs, that are shorter than 4B panics.

Ideally no utility should ever panic, so let's fix that first in this diff and report what has happened.

I'm also ignoring symrefs which were the problem in the first place.

Reviewed By: RajivTS

Differential Revision: D69521625

fbshipit-source-id: 8106da9cd00b6bc801cab4e6f347e0fcee7b4256