logging: replace slog with tracing_slog_compat
Summary:
We are migrating from slog to tracing.

Replace all dependencies on `slog` with a named depedency on `tracing_slog_compat`.  We continue to use slog, but this make it possible to start using tracing in upcoming diffs.

Mostly this is a drop-in replacement.  In a few places we must use `.into()` to convert real slog loggers to the tracing_slog_compat shim.

Reviewed By: RajivTS

Differential Revision: D68963022

fbshipit-source-id: f827773f97fe1850b9e505cc137739e35eddd527