Update system notifications with error summaries and remediation guidance
Summary:
Due to character limits in Windows system-level notifications, detailed error messages from health-report are getting truncated, making it difficult for users to understand the issues with their Eden setup.

To address this, we are implementing a map that provides brief summaries of errors and their corresponding remediation steps. This will enable users to quickly identify and resolve issues using the available diagnostic tools (E-menu and F-menu) thereby aligning with our goal of concise issue awareness and resolution.

Reviewed By: davidair, jdelliot

Differential Revision: D69506148

fbshipit-source-id: d516fd706d18d5a3c5078390df2b7c2ab41a5e5d