Define methods of privHelperConn to [de] serialize FAM requests and responses
Summary:
## This Diff

Defined methods of privConn to facilitate serializing and deserializing the requests to start/stop FAM.

## Context
File Access Monitor(FAM) is a tool to monitor file system access, aiming to find out EdenFS crawlers. The tool will be implemented after the toolchain is ready. Before that, we are going to use [this third-party binary](https://github.com/objective-see/FileMonitor).
This diff stack focuses on integrating the binary into Eden package and spawn a process from privHelper because of the required privileged permission.

Reviewed By: jdelliot

Differential Revision: D69340305

fbshipit-source-id: bfc78ccea0060ed47103b454cefc632dd47a2666