Define methods of privHelperServer to process FAM requests
Summary:
## This Diff
Defined two methods of privHelperServer to process FAM requests.

Fake data is returned for now.

## Context
File Access Monitor(FAM) is a tool to monitor file system access, aiming to find out EdenFS crawlers. The tool will be implemented after the toolchain is ready. Before that, we are going to use [this third-party binary](https://github.com/objective-see/FileMonitor).
This diff stack focuses on integrating the binary into Eden package and spawn a process from privHelper because of the required privileged permission.

Reviewed By: jdelliot

Differential Revision: D69340301

fbshipit-source-id: 07cf58df3a3da64662d7548f8b6796dc6cf48b43