fix a name change in code comments
Summary: The function previously known as `flushInvalidations` has been renamed to `completeInvalidations` in D44657395 for a better abstraction. This diff only updates the code comments to reflect this change for clarity. Although we have `EdenMount::flushInvalidations()`, in the context of these filesystem header files, we are specifically referring to `completeInvalidations`.

Reviewed By: MichaelCuevas

Differential Revision: D69667247

fbshipit-source-id: b39bfcc4f962175a8631f86d533359c642b92720