Implement concurrent fetching of pre-requisite objects during packfile processing
Summary:
One of the most time consuming phases of packfile processing during Git push is the fetching of pre-requisites objects required to resolve ref-deltas that are present in the packfile.

This bottleneck exists because we are fetching the pre-requisite objects one-by-one using `try_collect`. To optimize for available concurrency, this diff adds the necessary buffering support for we can fetch multiple pre-requisite objects in one go.

This should shave off valuable time for large pushes since for pushes of 1000 commits this routine takes ~40 seconds and for pushes of 10,000 commits this routines takes ~10 minutes.

Differential Revision: D69660657

fbshipit-source-id: 96bdefbfe1b98583660c42b3173f0c54d2492472