Enhance push-time perf telemetry for Mononoke Git Scuba
Summary:
Although we instrument every single (meaningful) phase of Git push on the server side, the telemetry logs in such cases just contain the tag, request_id, and timing information. This makes it difficult to fetch these logs for a particular user, repo or method.

Additionally, since we use `log_tags` for much more than push phase telemetry data, drilling down to the relevant logs requires hard coding which specific events you are interested in. This is tiresome and the user would likely give up by that time.

This diff makes the necessary changes so the push time telemetry logs are higher fidelity and easier to drill down to using approriate `msg` filter

Differential Revision: D69659533

fbshipit-source-id: 24dbc305d097947b0734b3e70a8fd81c857682d0