Update CLI to integrate thrift call
Summary:
## Context
File Access Monitor(FAM) is a tool to monitor file system access, aiming to find out EdenFS crawlers. The tool will be implemented after the toolchain is ready. Before that, we are going to use [this third-party binary](https://github.com/objective-see/FileMonitor).
This diff stack focuses on integrating the binary into Eden package and spawn a process from privHelper because of the required privileged permission.
## This Diff
Hook CLI commands with thrift calls.
## Next Step
Spawn a process and make it managed by privHelper server.

Reviewed By: jdelliot

Differential Revision: D68738598

fbshipit-source-id: f0bbf3a08dec0e98a1aa35bd31d67046d739b9e5