Improve security with conditional building
Summary:
## This Diff

Disable FAM completely for non-mac builds.

## Context
File Access Monitor(FAM) is a tool to monitor file system access, aiming to find out EdenFS crawlers. The tool will be implemented after the toolchain is ready. Before that, we are going to use [this third-party binary](https://github.com/objective-see/FileMonitor).
This diff stack focuses on integrating the binary into Eden package and spawn a process from privHelper because of the required privileged permission.

Reviewed By: ViniGupta08

Differential Revision: D69079397

fbshipit-source-id: ede66ccaa4dc2b15082d098950a6701f5ae6f859