Reuse preloaded draft changesets during bookmark movement for commit logging
Summary:
In D57906014, we streamified the loading of changesets for which hooks and checks need to be run. This was to ensure that we do not hold too many changesets in memory which was a concern for `gitimport` of large repos (like `chromium/src`).

However, as D66977548 demonstrated we eventually end up loading all the draft changesets that need to be logged in a vec of BonsaiChangesets. This is the same set for which we ran hooks in batches using a stream. Given that we eventually end up loading the changesets anyway, let's re-use the hook exeuction part of the pipeline to provide us with the needed vec of changesets.

However, instead of loading all the BonsaiChangesets in memory, let's load the needed `CommitInfo` objects which have a much smaller memory footprint. This allows the logging code to make use of pre-loaded CommitInfos instead of calling `find_draft_ancestors` which avoid repeated I/O work. In case of 1000 commit pushes, this makes a noticeable performance difference.

Differential Revision: D69744590

fbshipit-source-id: 77bee2b4b7c28470e49a1735a5d071594f958682