Define BookmarkInfoData in terms of CommitInfo instead of BonsaiChangeset
Summary:
The `BookmarkInfoData` struct is meant to represent the information needed for logging of changesets and bookmark movement after the bookmark create/update operation has been completed. Currently, its defined in terms of `BonsaiChangeset` for convenience but the actual logging happens with `CommitInfo`.

This diff directly uses `CommitInfo` within `BookmarkInfoData` and updates all the call sites accordingly. On its own, this diff is just a refactor, but the follow up diff will add more context to this change.

Differential Revision: D69744394

fbshipit-source-id: 4beb3c18fd7d608de81a2f2a96c79b268fdfddc3