mercurial_derivation: implement support for subtree changes
Summary:
**Introduce subtree operations to provide server-side support for efficient directory branching.**

This diff implements derivation of mercurial changesets when subtree operations are used.

For copy operations we must re-use the source mercurial manifest so that we match client behaviour.

If copy operations are present then the list of changed files in the commit text is expected to be empty.

Finally, we must generate an appropriate subtree commit extra for the subtree changes.

Reviewed By: mitrandir77

Differential Revision: D66499858

fbshipit-source-id: c6a44595aff4fa9be88a7fb51457d77ead424315