unodes: implement subtree copy support
Summary:
**Introduce subtree operations to provide server-side support for efficient directory branching.**

In this diff we implement support for deriving unodes, deleted manifests, and fastlog with subtree operations.

Unodes are required by design to be unique in the repository.  This means we cannot take advantage of the subtree copy to do O(1) derivation.  In the future we will replace unodes with a derived data type that understands subtree copies and won't have this restriction.

Similarly, deleted manifests must ignore the subtree operation as they do not have support in their design for a different history.  We base the deleted manifest on the unode tree so that the subtree operation is ignored.

Reviewed By: zzl0

Differential Revision: D66775635

fbshipit-source-id: 1a93ad993dbeb412cb5ece903999bbb612369145