pushrebase: switch to MPath for conflict detection
Summary:
With subtree operations, it now becomes possible for pushrebase conflicts to occur at the root path.  This is actually fairly unlikely, but it could happen if, for example, we are copying from the root of an old commit.

Change conflict detection (and conflict errors) to use `MPath` instead of `NonRootMPath` so that the root directory can be included in the checks.

Reviewed By: RajivTS

Differential Revision: D66499859

fbshipit-source-id: 65a536242254d4333032d6b5bf79d81a4030c390