commit_rewriting: allow rewrite of subtree copies that do not alter manifest
Summary:
When rewriting commits, we can accept rewrites of commits that include subtree operations as long as a subtree copy that re-uses the source tree isn't present.  We need to use a deep copy instead for these locations so that the commit transformation can work.

When we do these transformations, the subtree copy or merge information is lost, as we cannot guarantee that there is a corresponding source location for the subtree operation in the destination repository (e.g. the directory could have been copied from outside the small repo's area).  We deliberately remove the subtree metadata and mark the commit as created via lossy transformation.

Reviewed By: andreacampi

Differential Revision: D68706269

fbshipit-source-id: a8ae0e78374ef63b7e677b8e4a20ef4e2658f613