blame: implement subtree copy support
Summary:
**Introduce subtree operations to provide server-side support for efficient directory branching.**

In this diff we add support for deriving blame for files that have been copied or merged via subtree copy or merge.

This introduces a slight change to the blame API for parents of changes.  Previously, parents would always be a real parent of the blamed changeset, so we could provide the parent index.  Now it is possible for a parent to be the subtree copy source.  We call these parents "replacement parents" and provide the full hash in this case.

This fact was already true for mutable blame, and was one of the ways in which mutable blame was incorrect.  We can now use this feature for mutable blame also.

Reviewed By: mitrandir77

Differential Revision: D66775638

fbshipit-source-id: 023a8099b5d4de8976d9e0692524eec0793d1214