manifest: add subtree change support to bonsai_diff
Summary:
**Introduce subtree operations to provide server-side support for efficient directory branching.**

In this diff we add support for subtree operations to the `bonsai_diff` function.

Bonsai diff in the presence of a subtree operation means that we replace the parent manifests at the point in the tree of the subtree operation.  This means comparisons are performed against the manifest of the subtree operation source, and only changes relative to that are generated as bonsai differences.

Reviewed By: mitrandir77

Differential Revision: D66496122

fbshipit-source-id: 54538ce37ced718102f94842dc948a8576762f7b