hooks: add block_xrepo_subtree_copy_overlap hook
Summary:
We are not planning to implement full handling of subtree copies for paths that are being synced between small and large repos in the megarepo infra.  This is tricky to do, as there is no guarantee that there is a suitable corresponding copy source in the target repo for the sync.

Instead, prevent these from being landed with a hook.  We can work around this issue for copies that need it by using a deep copy instead.

In the future we may implement support, at which point this restriction can be relaxed.

Reviewed By: andreacampi

Differential Revision: D68705199

fbshipit-source-id: 13cd500eae820092dbecbf42a9b76b3edc3ab9bd