history_traversal: fix mutable blame parent identity
Summary:
When mutable blames are being generated, we used to lie about the parent index as there was no way to list the mutable blame source as a replacement parent.

Now that subtree operations have added that capability, make use of it to give more accurate mutable-history blames as well.

Reviewed By: andreacampi

Differential Revision: D68736391

fbshipit-source-id: 2d7871560b30cb0de3d6dcd6348c806e36da8efe