bssm3: implement subtree copy support
Summary:
**Introduce subtree operations to provide server-side support for efficient directory branching.**

In this diff we implement support for deriving bssm3 for commits that contain subtree operations.  Because of the nature of basename suffix skeleton manifests, there is no efficiency advantage when deriving them for subtree operations, so we must expand the subtree operation out to its full changes in order to use the existing code path to update the bssm.

Reviewed By: andreacampi

Differential Revision: D66837799

fbshipit-source-id: 122df67c8a9a3e8b45f2e4d59279adc2af06b9db