commit_graph: add subtree source edges
Summary:
Subtree operations (copies or merges) create a new kind of edge between commits.  For most commit graph operations these can be ignored, and this is what maintains the simplicity of the commit graph for most uses.

However, for derived data, we *do* potentially need the subtree source to be derived.  This means commit graph traversals for data derivation need to take subtree sources into account.

Most of the time, the subtree source is an ancestor or public commit, which means it will usually be the case that the data is already derived anyway.  The exception to this will be in backfilling processes, which need to take into account the whole shape of the graph.

To allow this, extend the commit graph to include additional edges for subtree sources.  Most of the time these will be ignored.  We will only use them for data derivation.

In this diff we add the defintion of subtree source edges and add them to the calculation process.  They are not yet used.

Reviewed By: andreacampi

Differential Revision: D68420963

fbshipit-source-id: 183ee849c5d3a415c06d3f5f7c68ef51c0926224